Release v2025.0.1

+++ include/dlstreamer/gst/mappers/gst_to_cpu.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2022 Intel Corporation
+ * Copyright (C) 2022-2025 Intel Corporation
  *
  * SPDX-License-Identifier: MIT
  ******************************************************************************/
@@ -85,7 +85,7 @@ class MemoryMapperGSTToCPU : public BaseMemoryMapper {
     FramePtr map_video(GstFramePtr src, AccessMode mode) {
         // Add GST_VIDEO_FRAME_MAP_FLAG_NO_REF to not increment/decrement GstBuffer ref-count during map/unmap
         GstMapFlags map_flags = mode_to_gst_map_flags(mode);
-        map_flags = static_cast<GstMapFlags>(map_flags | GST_VIDEO_FRAME_MAP_FLAG_NO_REF);
+        map_flags = static_cast<GstMapFlags>((unsigned)map_flags | (unsigned)GST_VIDEO_FRAME_MAP_FLAG_NO_REF);
 
         auto frame_ptr = std::make_shared<GstVideoFrame>();
         GstVideoFrame *frame = frame_ptr.get();
